+++ RepRep/RepRepArticleCollectionViewCell.swift
@@ -13,18 +13,8 @@ class RepRepArticleCollectionViewCell: UICollectionViewCell {
     
     static let id = "RepRepArticleCell"
     
-    var articleImageView: UIImageView = {
-        let view = UIImageView()
-        return view
-    }()
-    
-    var articleHeadlineLabel: UILabel = {
-        let view = UILabel()
-        view.numberOfLines = 0
-        view.textAlignment = .center
-        view.lineBreakMode = .byWordWrapping
-        return view
-    }()
+    var articleImageView = UIImageView()
+    var articleHeadlineLabel = RepRepLabel(type: .collectionCellLabel)
     
     var article: Article! {
         didSet {
@@ -32,10 +22,14 @@ class RepRepArticleCollectionViewCell: UICollectionViewCell {
         }
     }
     
+    override func prepareForReuse() {
+        articleImageView.image = nil
+    }
+    
     override init(frame: CGRect) {
         super.init(frame: frame)
-        setupViewHierarchy()
         configureConstraints()
+        articleImageView.contentMode = .scaleToFill
     }
     
     required init?(coder aDecoder: NSCoder) {
@@ -43,39 +37,26 @@ class RepRepArticleCollectionViewCell: UICollectionViewCell {
     }
     
     private func inputArticle() {
-        self.articleHeadlineLabel.text = self.article.headline
-        self.articleHeadlineLabel.alpha = 0.7
-        self.articleHeadlineLabel.font = UIFont.systemFont(ofSize: UIFont.systemFontSize)
-        self.articleHeadlineLabel.backgroundColor = .black
-        self.articleHeadlineLabel.textColor = UIColor.repCream
-        self.articleImageView.contentMode = .scaleAspectFit
-        if let url = self.article.thumbURL {
-            APIRequestManager.manager.getImage(APIEndpoint: "https://static01.nyt.com/\(url)") {(data) in
-                if let validData = data {
-                    DispatchQueue.main.async {
-                        self.articleImageView.image = UIImage(data: validData)
-                    }
-                }
-            }
-        }
-        //self.articleImageView.image
+        articleHeadlineLabel.text = article.headline
+        let gradient = CAGradientLayer()
+        gradient.colors = [UIColor.repGrey.withAlphaComponent(0), UIColor.repGrey]
+        articleHeadlineLabel.layer.addSublayer(gradient)
+        articleHeadlineLabel.layoutIfNeeded()
     }
     
     //MARK: - View Hierarchy and Constraints
     
-    func setupViewHierarchy () {
-        self.contentView.addSubview(articleImageView)
-        self.contentView.addSubview(articleHeadlineLabel)
-    }
-    
     func configureConstraints () {
+        contentView.addSubview(articleImageView)
+        contentView.addSubview(articleHeadlineLabel)
+        
         articleImageView.snp.makeConstraints { (view) in
-            view.trailing.leading.top.bottom.equalToSuperview()
+            view.top.bottom.equalToSuperview()
+            view.width.equalTo(UIConstants.collectionCellWidth)
         }
         
         articleHeadlineLabel.snp.makeConstraints { (view) in
-            view.leading.trailing.equalToSuperview()
-            view.centerY.equalToSuperview()
+            view.leading.trailing.bottom.equalToSuperview()
         }
     }
