+++ RepRep/RepRepCollectionView.swift
@@ -11,16 +11,18 @@ import UIKit
 class RepRepCollectionView: UICollectionView {
     init() {
         let layout = UICollectionViewFlowLayout()
-        layout.itemSize = UICollectionViewFlowLayoutAutomaticSize
+        //layout.itemSize = UICollectionViewFlowLayoutAutomaticSize
         layout.sectionInset = UIEdgeInsets(top: 0,
-                                           left: CGFloat(UIConstants.innerBorderInset),
-                                           bottom: 0,
-                                           right: CGFloat(UIConstants.innerBorderInset))
-        layout.scrollDirection = .horizontal
-        layout.estimatedItemSize = .zero
+                                           left: 0,
+                                           bottom: CGFloat(UIConstants.innerBorderInset),
+                                           right: 0)
+        layout.scrollDirection = .vertical
+        layout.itemSize = CGSize(width: UIConstants.collectionCellWidth, height: 80)
         super.init(frame: .zero, collectionViewLayout: layout)
         register(RepRepArticleCollectionViewCell.self, forCellWithReuseIdentifier: RepRepArticleCollectionViewCell.id)
         backgroundColor = UIColor.repCream
+        showsVerticalScrollIndicator = false
+        showsHorizontalScrollIndicator = false
     }
     
     required init?(coder aDecoder: NSCoder) {
