+++ RepRep/UIConstants.swift
@@ -1,0 +1,19 @@
+//
+//  UIConstants.swift
+//  RepRep
+//
+//  Created by Liam Kane on 8/30/17.
+//  Copyright © 2017 RepresentativeReputation. All rights reserved.
+//
+
+import UIKit
+
+struct UIConstants {
+    static let innerBorderInset = 24
+    static let outerBorderInset = 44
+    static let titleFontSize: CGFloat = 20
+    static let roundCornerValue: CGFloat = 20
+    static let partyIconHeightHeight: CGFloat = 40
+    static let statusBarSize: CGFloat = 20
+    static let collectionCellWidth = UIScreen.main.bounds.width / 2 - CGFloat(UIConstants.innerBorderInset + 12)
+}
